Refactor: JobZonePage into smaller components
Refactor JobZonePage.tsx into smaller, focused components, including JobZoneHeader, JobDescriptionInput, GenerationProgress, ResumeDisplay, and ResumeActions.
