+++ src/components/job-parser/ResumeExportHandler.tsx
@@ -1,0 +1,92 @@
+
+import React from 'react';
+import { toast } from '@/hooks/use-toast';
+import { parseHTMLToPDFSections, generatePDFFromSections } from '@/utils/pdfGenerator';
+
+export const useResumeExport = () => {
+  const downloadPDF = async () => {
+    const resumeElement = document.getElementById('resumeOutput');
+    
+    if (!resumeElement) {
+      toast({
+        title: 'Error',
+        description: 'Resume content not found. Please generate a resume first.',
+        variant: 'destructive',
+      });
+      return;
+    }
+
+    if (!resumeElement.innerHTML.trim() || resumeElement.innerHTML.trim() === '') {
+      toast({
+        title: 'Error',
+        description: 'No resume content to generate PDF. Please generate a resume first.',
+        variant: 'destructive',
+      });
+      return;
+    }
+
+    try {
+      toast({
+        title: 'Generating PDF',
+        description: 'Please wait while we create your PDF...',
+      });
+
+      const sections = parseHTMLToPDFSections(resumeElement.innerHTML);
+      
+      if (sections.length === 0) {
+        throw new Error('No content could be extracted from the resume');
+      }
+
+      generatePDFFromSections(sections, 'resume.pdf');
+      
+      toast({
+        title: 'Success',
+        description: 'PDF downloaded successfully!',
+      });
+    } catch (error) {
+      console.error('Error generating PDF:', error);
+      toast({
+        title: 'Error',
+        description: 'Failed to generate PDF. Please try the copy option instead.',
+        variant: 'destructive',
+      });
+    }
+  };
+
+  const copyToClipboard = async () => {
+    const resumeElement = document.getElementById('resumeOutput');
+    
+    if (!resumeElement) {
+      toast({
+        title: 'Error',
+        description: 'Resume content not found.',
+        variant: 'destructive',
+      });
+      return;
+    }
+
+    try {
+      const plainText = resumeElement.innerText || resumeElement.textContent || '';
+      
+      if (!plainText.trim()) {
+        throw new Error('No text content found');
+      }
+
+      await navigator.clipboard.writeText(plainText);
+      
+      toast({
+        title: 'Success',
+        description: 'Resume content copied to clipboard!',
+      });
+    } catch (error) {
+      console.error('Error copying to clipboard:', error);
+      toast({
+        title: 'Error',
+        description: 'Failed to copy content to clipboard.',
+        variant: 'destructive',
+      });
+    }
+  };
+
+  return { downloadPDF, copyToClipboard };
+};
